Restructure into extension
