+++ YT Music/Supporting Files/Assets.xcassets/AppIcon.appiconset/Icon.png
