+++ YT Music/Supporting Files/Assets.xcassets/AppIcon.appiconset/icon_32x32@2x.png
