+++ YT Music/Supporting Files/Assets.xcassets/Back Arrow Active.imageset/Back Arrow Active.pdf
