+++ YT Music/Supporting Files/Assets.xcassets/Contents.json
